requirements.txt
flask-admin
Flask-Bcrypt
Flask-Injector
flask-migrate
flask-restx
flask-sqlalchemy
numpy
opencv-python
psycopg2-binary
pytesseract
python-dotenv

